Change command-line args.
I changed the command-line args to
* day number: on which day of the meet the script is being run
* platform ID: the platform of interest
* output file path: optional, the file into which the script will write the data
  to be displayed by DRL

Each day of the meet will be a meet in liftingcast.com and therefore in CouchDB.
The script will set the meet ID and password (for accessing the correct db)
based on the day number provided by the user.

Note that I've hard-coded the 2018 Aggie BBQ meet ID and password for now. The
above function is to be written.