Sort attempt-lifter tuples into lifter order.
We fetch lifters for the given platform and all attempts in the meet from the
db. We inner join attempts onto lifters on lifter ID, ending up with
attempt-lifter tuples, which we then sort into the lifting order.
This is the lifting order for the platform for the entire day of the "meet",
where the "meet" database has the data for one day of the real-life meet.

Note that the lifter order excludes attempts without a weight field or with
non-positive weight:
* Attempts that have not yet been submitted will not be in the lifter order,
  though liftingcast.com will have created the respective documents for them.
* Attempts that will be passed get assigned 0 weight in liftingcast.com and so
  will not be in the lifting order.