Fix lifter ID lookup to avoid key error.
Given an attempt for some lifter in the meet, I get the lifter with the
attempt's "lifterId" from the lifters_by_id dictionary of lifters FOR OUR
PLATFORM. The attempt may be for a lifter on another platform, in which case
the attempt's "lifterId" will not be a key of the dictionary. That's fine and
expected so it should not raise a KeyError.