Improve message for missing dependencies.
Note that enum must be installed for Python 2 so I moved it to the modules that
we try to import.
Normally, the stacktrace would be fine -- there would be no reason to try-except
imports. If something goes wrong and time is a factor, though, then I want
anyone who ran the script to be able to fix the problem.